requirements.txt
requests==2.32.5
Flask==3.1.2
sentence-transformers==5.1.1
pinecone-client==4.0.0
chromadb==1.1.0
pandas==2.3.2
matplotlib==3.10.6
seaborn==0.13.2
ollama==0.6.0
python-dotenv==1.1.1
ipykernel==6.30.1
pydantic==2.11.9
prometheus-client==0.23.1
pytest==8.4.2
rapidfuzz==3.14.1
instructor==1.3.3
openai==1.35.3
